gui/requirements_gui.txt
# Requirements supplémentaires pour l'interface graphique
Pillow>=9.0.0
